move ErrorCard to package-card from ConverterScreen
